Add first orb version
After a bit of trial-and-error, I landed in this configuration. There is
one command, `run_happo`. It will trigger a happo CI run in the current
directory.

There's also a job, `happo/run` that puts together all necessary parts
to run in isolation. With this job, users' circleci configs can look as
simple as

```yaml
version: 2.1

orbs:
  happo: happo/happo@dev:seventh

workflows:
  version: 2.1
  run_all:
    jobs:
      - happo/run
```